warning and compilation correction

+++ MathPhys/main.cpp
@@ -2,7 +2,7 @@
 #include <GL/gl.h>
 #include <GL/glu.h>
 #include <iostream>
-#include "Particule.h"
+#include "Particle.h"
 #include "Vecteur3D.h"
 
 using namespace std;
@@ -13,7 +13,7 @@ float sphereObjZ = 0;
 
 Particle projectile;
 
-float frameTime = 0.0033;
+float frameTime = 0.0033f;
 
 float MatSpec[4] = { 0.1f, 0.1f, 0.5f, 1.0f };
 float MatDif[4] = { 0.057f, 0.447f, 0.361f, 1.0f };
@@ -40,9 +40,9 @@ void translation(Vecteur3D vec1)
 {
 	projectile.integrate(frameTime);
 
-	sphereObjX = sphereObjX + vec1.getx();
-	sphereObjY = sphereObjY + vec1.gety();
-	sphereObjZ = sphereObjZ + vec1.getz();
+	sphereObjX = sphereObjX + static_cast<float>(vec1.getx());
+	sphereObjY = sphereObjY + static_cast<float>(vec1.gety());
+	sphereObjZ = sphereObjZ + static_cast<float>(vec1.getz());
 
 }
 
@@ -73,7 +73,7 @@ void displayLoop(void)
 
 	Vecteur3D vec1 = projectile.getPosition();
 	glTranslatef(sphereObjX - 2, sphereObjY, sphereObjZ);
-	initSphereObj(0.1);
+	initSphereObj(0.1f);
 	glPopMatrix();
 
 	translation(vec1);
